Version: 2003-12-15 ksh93p-
Sourced from ast-open-history version 2003-09-23 and:
eeefe1456a71511be041d0c8f5f82e44b79f572a61b13554408d835b91379119  INIT.2003-12-15.tgz
1752a9d917243bd4c598438bf643408d9c315e56925b560568432c06f30457a9  ast-base.2003-12-15.tgz

+++ src/lib/libast/regex/reglib.h
@@ -122,7 +122,7 @@ typedef struct regsubop_s
 #define REX_WORD		42	/* word boundary		*/
 #define REX_WORD_NOT		43	/* not word boundary		*/
 
-#define T_META		(UCHAR_MAX+1)
+#define T_META		((int)UCHAR_MAX+1)
 #define T_STAR		(T_META+0)
 #define T_PLUS		(T_META+1)
 #define T_QUES		(T_META+2)
